backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
pymysql==1.1.1
cryptography==43.0.1
redis==5.1.1
rq==1.16.2
bcrypt==4.2.1
PyJWT==2.9.0
pydantic-settings==2.5.2
python-multipart==0.0.12
httpx==0.27.2
pytest==8.3.3
